+++ apps/storybook/stories/landing-page.stories.tsx
@@ -1,0 +1,36 @@
+import { fn } from "@storybook/test";
+import LandingPage from "./components/landingPage";
+
+export default {
+  title: "@govtechmy/examples/LandingPage",
+  component: LandingPage,
+  // remove the automatic documentation for Layout
+  tags: ["!autodocs"],
+  parameters: {
+    layout: "fullscreen",
+  },
+};
+
+export const Mobile = {
+  parameters: {
+    viewport: {
+      defaultViewport: "mobile1",
+    },
+  },
+};
+
+export const Tablet = {
+  parameters: {
+    viewport: {
+      defaultViewport: "tablet",
+    },
+  },
+};
+
+export const Desktop = {
+  parameters: {
+    viewport: {
+      defaultViewport: "desktop",
+    },
+  },
+};