+++ packages/react/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @govtechmy/myds-react
 
+## 0.0.21
+
+### Patch Changes
+
+- e71f6dc: Move eslint config from dependency to devDependency
+
+## 0.0.20
+
+### Patch Changes
+
+- d92931f: Implement shared eslint config, and fixed lint issues
+- d897cd3: Fix navbar height for mobile
+- Updated dependencies [d92931f]
+  - @myds/eslint-config@0.0.1
+
+## 0.0.19
+
+### Patch Changes
+
+- 02810f0: Support asChild interaction for NavbarMenuItem
+- fe0b3be: Update color of table cell
+- b73a362: Fix Tabs line style off by 1px and remove divider
+
 ## 0.0.18
 
 ### Patch Changes